requirements.txt
langchain==0.3.27
langchain-text-splitters==0.3.11
langchain-google-genai==2.1.10
langchain-community==0.3.29
langchain-chroma==0.2.4
langgraph==0.6.4
chromadb==1.0.20
pypdf==6.0.0
shiny
python-dotenv==1.1.1
google-genai
rsconnect-python==1.28.2
